requirements.txt
PyYAML
scikit-learn
nltk
spacy
networkx